app/requirements.txt

fastapi
uvicorn
nats-py
weaviate-client
openai
requests
sqlalchemy
psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
aiohttp
